# model/model_notification.go
/*
 * Demo App
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Notification struct {

	Id int32 `json:"id,omitempty"`

	Priority int32 `json:"priority,omitempty"`

	Type_ int32 `json:"type,omitempty"`

	Title string `json:"title,omitempty"`

	Content string `json:"content,omitempty"`

	PosterID int32 `json:"posterID,omitempty"`

	DateCreated string `json:"dateCreated,omitempty"`

	DateExpired string `json:"dateExpired,omitempty"`
}